backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.2
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
djangorestframework==3.16.1
frozenlist==1.7.0
geographiclib==2.1
geopy==2.4.1
idna==3.10
kombu==5.5.4
multidict==6.6.4
new-package==0.0.1
numpy==2.3.3
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
pyarrow==21.0.0
PyJWT==2.10.1
pylance==0.37.0
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==8.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
twilio==9.8.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
yarl==1.20.1
